feat: only sponsor creates winner banners
