+++ src/features/listings/components/ListingPage/ListingWinners.tsx
@@ -1,27 +1,14 @@
-import {
-  Box,
-  Button,
-  Center,
-  Flex,
-  Image,
-  Text,
-  useToast,
-} from '@chakra-ui/react';
+import { Box, Button, Center, Flex, Image, Text } from '@chakra-ui/react';
 import axios from 'axios';
 import Avatar from 'boring-avatars';
-import html2canvas from 'html2canvas';
 import NextLink from 'next/link';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import type { SubmissionWithUser } from '@/interface/submission';
-import { getBlobFromCanvas } from '@/utils/canvasToBlob';
-import { openExternalLinkInNewTab } from '@/utils/linkInNewTab';
 import { sortRank } from '@/utils/rank';
 import { tweetEmbedLink, tweetTemplate } from '@/utils/tweetTemplate';
-import { uploadToCloudinary } from '@/utils/upload';
 
 import type { Bounty, Rewards } from '../../types';
-import WinnerBanner from './WinnerBanner';
 
 interface Props {
   bounty: Bounty;
@@ -30,15 +17,8 @@ interface Props {
 export function ListingWinners({ bounty }: Props) {
   const [isListingLoading, setIsListingLoading] = useState(true);
   const [submissions, setSubmissions] = useState<SubmissionWithUser[]>([]);
-  const [loadingBanner, setLoadingBanner] = useState(false);
-  const [bannerUrl, setBannerUrl] = useState(bounty.winnerBannerUrl);
-
-  const winnerBannerRef = useRef<HTMLDivElement>(null);
-  const shareButtonRef = useRef<HTMLButtonElement>(null);
 
   const isProject = bounty?.type === 'project';
-
-  const toast = useToast();
 
   const getSubmissions = async (id?: string) => {
     setIsListingLoading(true);
@@ -66,119 +46,12 @@ export function ListingWinners({ bounty }: Props) {
     getSubmissions();
   }, []);
 
-  const saveBannerUrl = async (): Promise<string> => {
-    return new Promise(async (resolve, reject) => {
-      try {
-        if (!winnerBannerRef.current) {
-          setLoadingBanner(false);
-          throw new Error('no banner ref');
-        }
-        console.log('oooo no banner url');
-        let endTime,
-          computeTime,
-          startTime = performance.now();
-        const canvas = await html2canvas(winnerBannerRef.current, {
-          useCORS: true,
-          width: 1200,
-          height: 675,
-          x: 0,
-          y: 0,
-          onclone: (el) => {
-            const elementsWithShiftedDownwardText =
-              el.querySelectorAll<HTMLElement>('.shifted-text');
-            elementsWithShiftedDownwardText.forEach((element) => {
-              element.style.transform = 'translateY(-30%)';
-            });
-          },
-        });
-        endTime = performance.now();
-        computeTime = endTime - startTime;
-        console.log(`Canvas Created: ${computeTime} milliseconds`);
-        startTime = performance.now();
-
-        const mimeType = 'image/png';
-        if (!bounty.id || !bounty.slug) throw new Error('no id or slug');
-        const fileName = `${bounty.id}-winner-banner`;
-
-        const blob = await getBlobFromCanvas(canvas, mimeType);
-
-        endTime = performance.now();
-        computeTime = endTime - startTime;
-        console.log(`Image Created: ${computeTime} milliseconds`);
-        startTime = performance.now();
-
-        if (!blob) throw new Error('no blob');
-        const file = new File([blob], fileName, { type: mimeType });
-        console.log('image size - ', blob.size, file.size);
-
-        // NEED THIS FOR LOCAL DOWNLOAD LINK, WE DONT WANT TO SPAM CLOUDINARY FOR IMAGE LINK
-        // const localUrl = URL.createObjectURL(file);
-        // const downloadLink = document.createElement('a');
-        // downloadLink.href = localUrl;
-        // downloadLink.setAttribute('download', 'data.png'); // Name the file here
-        // document.body.appendChild(downloadLink);
-        // downloadLink.click();
-        // document.body.removeChild(downloadLink);
-        // throw new Error("done")
-
-        const url = await uploadToCloudinary(file);
-        setBannerUrl(url);
-
-        endTime = performance.now();
-        computeTime = endTime - startTime;
-        console.log(`Uploaded to Cloudinary: ${computeTime} milliseconds`);
-        startTime = performance.now();
-
-        await axios.put(`/api/bounties/${bounty.slug}/setWinnerBanner`, {
-          image: url,
-        });
-
-        endTime = performance.now();
-        computeTime = endTime - startTime;
-        console.log(`Updated Database: ${computeTime} milliseconds`);
-        startTime = performance.now();
-
-        resolve(url);
-      } catch {
-        reject('Some Error Occured');
-        setLoadingBanner(false);
-        toast({
-          title: 'Failed to Share',
-          description: 'Please try again later',
-          status: 'error',
-          duration: 5000,
-          isClosable: true,
-          variant: 'subtle',
-        });
-      }
-    });
-  };
-
   const openWinnerLink = () => {
-    console.log('we got banner url');
     let path = window.location.href.split('?')[0];
-    if (!path) {
-      setLoadingBanner(false);
-      return;
-    }
-
+    if (!path) return;
     path += 'winner/';
 
-    const tweetLink = tweetEmbedLink(tweetTemplate(path));
-
-    setLoadingBanner(false);
-    openExternalLinkInNewTab(tweetLink);
-  };
-
-  const onShareClick = async (bannerUrl: string | undefined) => {
-    setLoadingBanner(true);
-    if (bannerUrl) {
-      openWinnerLink();
-      return;
-    }
-    saveBannerUrl();
-    openWinnerLink();
-    setLoadingBanner(false);
+    return tweetEmbedLink(tweetTemplate(path));
   };
 
   if (isListingLoading || !submissions.length) {
@@ -196,16 +69,6 @@ export function ListingWinners({ bounty }: Props) {
       >
         🎉 Winners Announced
       </Text>
-      {!bannerUrl && (
-        // <Box pos="fixed" zIndex={1} top={'10%'} right={'30%'}>
-        <Box pos="fixed" zIndex={-99999} top={'-300%'} right={'-300%'}>
-          <WinnerBanner
-            ref={winnerBannerRef}
-            bounty={bounty}
-            submissions={submissions}
-          />
-        </Box>
-      )}
       <Box mx={3}>
         <Box
           pos="relative"
@@ -285,42 +148,39 @@ export function ListingWinners({ bounty }: Props) {
               </NextLink>
             ))}
           </Flex>
-          <Button
-            ref={shareButtonRef}
-            pos="absolute"
-            top={5}
-            right={5}
-            gap={2}
-            display="flex"
-            color="rgba(0, 0, 0, 0.65)"
-            fontSize="14px"
-            fontWeight={500}
-            bg="white"
-            _hover={{ background: 'rgba(255, 255, 255, 0.8)' }}
-            _active={{ background: 'rgba(255, 255, 255, 0.5)' }}
-            isLoading={loadingBanner}
-            onClick={() => onShareClick(bannerUrl)}
-          >
-            Share on
-            <Center w="1.2rem">
-              <svg
-                width="33px"
-                height="33px"
-                viewBox="0 0 33 33"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M25.0851 3.09375H29.6355L19.6968 14.4504L31.3886 29.9062H22.2363L15.0626 20.5348L6.86421 29.9062H2.30737L12.9357 17.7568L1.72729 3.09375H11.1117L17.5892 11.6596L25.0851 3.09375ZM23.4867 27.1863H26.0068L9.73882 5.67188H7.03179L23.4867 27.1863Z"
-                  fill="black"
-                />
-              </svg>
-            </Center>
-          </Button>
+          <NextLink href={openWinnerLink() ?? '#'} target="_blank">
+            <Button
+              pos="absolute"
+              top={5}
+              right={5}
+              gap={2}
+              display="flex"
+              color="rgba(0, 0, 0, 0.65)"
+              fontSize="14px"
+              fontWeight={500}
+              bg="white"
+              _hover={{ background: 'rgba(255, 255, 255, 0.8)' }}
+              _active={{ background: 'rgba(255, 255, 255, 0.5)' }}
+            >
+              Share on
+              <Center w="1.2rem">
+                <svg
+                  width="33px"
+                  height="33px"
+                  viewBox="0 0 33 33"
+                  fill="none"
+                  xmlns="http://www.w3.org/2000/svg"
+                >
+                  <path
+                    d="M25.0851 3.09375H29.6355L19.6968 14.4504L31.3886 29.9062H22.2363L15.0626 20.5348L6.86421 29.9062H2.30737L12.9357 17.7568L1.72729 3.09375H11.1117L17.5892 11.6596L25.0851 3.09375ZM23.4867 27.1863H26.0068L9.73882 5.67188H7.03179L23.4867 27.1863Z"
+                    fill="black"
+                  />
+                </svg>
+              </Center>
+            </Button>
+          </NextLink>
         </Box>
       </Box>
     </Box>
   );
 }
-
-// <NextLink href={`/listings/${bounty?.type}/${bounty?.slug}/winners/`}>
