+++ src/features/sponsor-dashboard/components/Submissions/index.ts
@@ -1,3 +1,4 @@
 export * from './SubmissionDetails';
 export * from './SubmissionHeader';
 export * from './SubmissionList';
+export * from './WinnerBanner';
